+++ Source/Krypton Components/Krypton.Toolkit/Rendering/KryptonMicrosoft365Renderer.cs
@@ -838,11 +838,6 @@ public class KryptonMicrosoft365Renderer : KryptonProfessionalRenderer
         {
             case ContextMenuStrip or ToolStripDropDownMenu:
             {
-                // Make sure the font is current
-                if (e.ToolStrip.Font != KCT.MenuStripFont)
-                {
-                    e.ToolStrip.Font = KCT.MenuStripFont;
-                }
 
                 // Create border and clipping paths
                 using GraphicsPath borderPath = CreateBorderPath(e.AffectedBounds, _cutContextMenu),
@@ -859,11 +854,6 @@ public class KryptonMicrosoft365Renderer : KryptonProfessionalRenderer
                 if (TryRenderStatusStripOverride(e, e.Graphics))
                 {
                     break;
-                }
-                // Make sure the font is current
-                if (e.ToolStrip.Font != KCT.StatusStripFont)
-                {
-                    e.ToolStrip.Font = KCT.StatusStripFont;
                 }
 
                 // We do not paint the top two pixel lines, as they are drawn by the status strip border render method
@@ -886,23 +876,14 @@ public class KryptonMicrosoft365Renderer : KryptonProfessionalRenderer
 
                 break;
             }
-            // Make sure the font is current
             case MenuStrip:
             {
-                if (e.ToolStrip.Font != KCT.MenuStripFont)
-                {
-                    e.ToolStrip.Font = KCT.MenuStripFont;
-                }
 
                 base.OnRenderToolStripBackground(e);
                 break;
             }
             default:
             {
-                if (e.ToolStrip.Font != KCT.ToolStripFont)
-                {
-                    e.ToolStrip.Font = KCT.ToolStripFont;
-                }
 
                 // Cannot paint a zero sized area
                 var backRect = new RectangleF(0, 0, e.ToolStrip.Width, e.ToolStrip.Height);